Resolved Hibernate issue with null collections.
Issue exception:
org.hibernate.HibernateException: A collection with cascade="all-delete-orphan" was no longer referenced by the owning entity instance

The following link provides some more info about the issue:
https://hibernate.atlassian.net/browse/HHH-9940

Deduced cause :
The orphan removal association's constraint being entirely an ORM concept, as the underlying DB has nothing to do wit it, on an entity update, to ensure this constraint, ORMs need to compare the managed entity object with the new object updated state. If the reference to a previously managed associated entity is modified, in the managed constraint-defining entity, the ORM proceeds to remove it from the DB.

For non one-to-one associations (many), Hibernate ensures this behavior by checking against the entries in the association's collection object.
If the association's collection reference changes for the managed entity with an orphan removal constraint, Hibernate throws the above mentioned exception.
Since in java (null == null) is not true and Hibernate apparently simply compares references (with the double equals sign operator), it doesn't make an exception for this case scenario. Thus seeing the pointed at collection reference (which is null) as different from itself, and proceeding to throw the exception even though it wasn't changed.